DOC: Update internal links for generator.rst and related
This is a first in a sequence of future PRs to update internal links, as discussed in the triage meeting.
In addition to removing double backticks on internal functions, I removed unneeded tilde ~ as well.
This is a great example page to see why a tilde ~ is needed in the first paragraph of generator.rst.

I included instructions for the POSSEE team on this topic, as its a simple way to help train them in the PR process.

[skip azp] [skip actions] [skip cirrus]