TST: Make pass <3.11
On older snakes, was previously assigning Array = NDArray.  However,
that was causing unexpected errors in testing the shape311.py files.
Had to extend numpy.typing.Array to <3.11 using typing_extensions, as
well as quoting some type hints in the test files for delayed
evaluation.

Alternatively, I could have found a way to limit the parameterization
of test_fail and test_runs for just >= 3.11.  However, IDK how to do
that in pytest.