BUG: Fix in1d fast-path range
The code never worked when the first array was uint and the second
one had negative values, NumPy 2 makes that slightly worse because
the reverse is also true (the second one is uint).

I did that by just using intp.  The indexing operation has to cast
anyway, so it seems unlikely that we have much of a downside in
general.
Casting there seems to make one bounds check just unnecessary,
so removed it.

(Yes, I guess indexing could use a buffered iterator so if both
ar1 and ar2 are huge and ar1[basic_mask] is so huge it barely
fits the memory, that is a  downside.  I don't think I care, but
if someone does it can be modified probably)