Use `cibuildwheel` for Pyodide/Emscripten CI job
This commit performs the following changes:

1. cibuildwheel version 2.19.1 is added for building
and testing the Pyodide wheels for NumPy in its workflow
for out-of-tree builds.
2. The workflow sets the CIBW_PLATFORM environment
variable in order to point to Pyodide as the build target.
3. The version of Pyodide used is bumped to the recent
version 0.26.1.
4. The TOML tables are updated to use cibuildwheel's
provided overrides for the test command, the wheels'
repair command, and the setup arguments needed at the
time of building the wheel.

[skip azp] [skip circle] [skip cirrus]