ENH: Add unstack()
unstack() is a new function from the 2023.12 version of the array API, which
serves as the inverse of stack(), that is, unstack(x, axis=axis) returns a
tuple of arrays along axis that when stacked along axis would recreate x.

The implementation is in pure Python, since it is just a straightforward
iteration along an index, but if it is preferable it can be moved to an
implementation in C.

I haven't yet added any tests.