DOC: Mention more error paths and try to consolidate import errors
The first section was nice, but outdated since NumPy 1.25 effectively
solved the issue so that builds have to be quite old to run into the
previous paths.

OTOH, the NumPy 2 specific section would now overlap if fixing the
first.  So tried to use the first section but expand/rephrase it.

In practice, C-API/Attribute errors, etc. won't be the only reason
for things not working (e.g. the error could just as well be a missing
attribute due to the Python API change), but let's focus on the C API...

My main thought was that mentioning that dtype change error is good,
it is an unfortunate thing that Cython seems to check that before
trying to import NumPy, but it is also a pretty "obvious" error.