DOC: Update roadmap a bit more
Just some thoughts, I didn't include ufunc chaining, although I think
it would also be valid to mention it (even if I doubt anyone will work
on it).  All others are things that have been discussed multiple times
at least, so I think they are fine to add, even if I don't know if
anyone will work on it:  That is true for practically everything.

I also always play with the idea that labeled axes might not be all that
bad to think about.  But it may need a quite a lot of dedication.
(It might not be that bad, but its a pretty broad thing.)