TST: Fix fp_noncontiguous and fpclass on riscv64
Modify fp_noncontiguous and fpclass so that they pass on riscv64.
The subtests that verify that the signs of negative NaNs are preserved
do not pass on riscv64 builds and are disabled.  Many RISC-V
instructions that produce NaNs return a canonical NaN, as defined by
the RISC-V specification.  The canonical NaNs are always positive. See
section 11.3 NaN Generation and Propagation of the RISC-V Unprivileged
ISA for more details.