BUG: allow replacement in the dispatch cache (#26693)
Allow identical replacement in dispatch cache, since this can be hit with freethreaded Python

* TST: add a failing test for dispatch cache thread safety

* TST: test replace=True error with new semantics

First version added a lock, but maybe some information is interesting:

The test code I added raises an unhandled thread exception every time I run it using a linux x86 laptop without the change to dispatching.c in this PR.

Still not sure why this failure is only hit under gcc.

Here's what I think is happening: right now the locking in the dispatch cache is only internal to the cache itself. The current locking strategy allows a race condition where two threads simultaneously see the cache is empty early in dispatch and then try fill it. The second thread to fill it sees the cache is filled and raises an exception, because the replace parameter for PyArrayIdentityHash_SetItem is 0, so replacements raise an exception.

I don't think it's possible to support this replace feature without moving the lock from the dispatch cache struct to somewhere in the dispatching logic in dispatching.c. We'd need to lock around all the spots we check for an entry in the dispatch cache and then later insert an entry into the cache. Happy to try that approach if it turns out replacing entries in this cache is problematic for some reason. I didn't want to do that since this code is hit every time a ufunc is called so I don't want to add even larger blocks of code that need to be locked around.

In practice, I don't think it's problematic to simply replace entries when this happens, at least not any more problematic than the current approach, since the dispatch cache holds borrowed references to ArrayMethod instances.

Fixes #26690.