TYP: fix incorrect import in `ma/extras.pyi` stub
The `numpy.ma` tests for type stubs are missing, so this kind
of obvious error has a chance of creeping in. The import in the
type stub now matches the one in the corresponding .py file again.

Also add the start of a very basic "pass" typing test for numpy.ma

[skip azp] [skip cirrus] [skip circle]