DOC: Change NEP hardlinks in NEP 35 to intersphinx mappings.
These changes are all in a footnote section at the end of NEP 35.

In addition to simply converting the links for NEPs 18, 30, and 37,
I also changed the target of the link for NEP 31. It has been linking
to pull request 14389 rather than to NEP 31 itself. I'm guessing the
two NEPs may have been in draft at the same time and the PR was the
best link available for NEP 31. But since 31 is published, it seems
reasonable to link to it rather than a draft PR for it.

This is another installment towards addressing #26707.

[skip actions] [skip azp] [skip cirrus]