ENH: Allow use of clip with Python integers to always succeed
This helps with NEP 50 behavior since the type is unchanged in that
case, but the result is still clearly defined for clipping.

This part could be pushed into the loops, but is more tricky to
do there.  It may make sense if it comes up as a problem with
minimum/maximum.  However, I think the situation there is slightly
less surprising.

(Does not preserve NEP 50 warnings/old behavior.)

Closes gh-26759