ENH: Provide a hook for gufuncs to process core dimensions.
The field `process_core_dims_func` is added to the PyUFuncObject.
This is a pointer to a function typedef'd as

    typedef int (PyUFunc_ProcessCoreDimsFunc)(
                                    struct _tagPyUFuncObject *ufunc,
                                    npy_intp *core_dim_sizes);

The author of a gufunc can set the field with a function that they
implement.  The function will be called when the gufunc is called.
(The actual call is in the internal function _get_coredim_sizes.)
The user-defined function can set an exception and return an error
status if any of the core dimensions in `core_dim_sizes` do not
satisfy the assumptions of the gufunc. The user-defined function
can also *set* the value of core dimensions that are passed in as -1,
meaning the correspond out parameter was not given.  This allows
calculations such pairwise distances (which generates m*(m-1)/2
output values for an input with shape (m, n)) and full convolution
(generates m + n - 1 output values from two inputs with shapes m
and n) to be implemented as gufuncs with automatic allocation of
the output with the correct shape.  The output shape is computed
and set in the user-defined function.