API: Partially revert unique with return_inverse
There was a good argument that it is not possible to reconstruct
the original array with `axis=None` without first reshaping and
changing the result shape helped with it.

However, it was always possible to do it for other axis values
by using `np.take` rather than `np.take_along_axis`.  Changing it
for all axis values is unnecessary to achieve reconstruction because
`np.take(arr, inverse, axis=axis)` already performed the job except
for `axis=None`.

Thus, this keeps the change for axis=None, but reverts gh-25553
for numerical axis.