BUG,MAINT: Fix utf-8 character stripping memory access
This fixes the memory access bug, the old if was there for a reason
(obviously...), but unfortunately only the sanitizer checks noticed
that.

But to make it clear, I had to just also rename/change things a bit
making it unsigned and using a stop range rather than the actual
range is just much clearer here where the range can go to length 0
IMO.

I.e. the old code had to check for `j >= 0` not `j > 0` (j being the
last character index), because `j` could go negative.

Fixes the sanitizer tests.