BUG: Fix out-of-bound minimum offset for in1d table method
We could always use ar2.dtype, but we want intp anyway, so that
is nicer.  But if the minimum value doesn't fit intp, NumPy would
complain (it may even guaranteed to work to force-cast,
but...)

So just try to use intp, but if that doesn't work use ar2.dtype
which must work, but means we'll have to cast once more (even if
that cast is probably just a view in principle).

Closes gh-26922