BUG: Use fused-multiply-add for complex operations (fixes for #26940 and and #26740)
* #26940 already fixed in same PR by nomemoverlap fix (failure depended on existence of two problems)
* For #26740 only consistency within numpy is fixed. Regular complex in Python may still have different results

Currently, fused-multipliy-add is typically used everywhere in numpy, depending on the compiler.
But some compilers may choose not to use fma,
as is the case in numpy-1.26.4 (and any <2 I believe) on macOS on ARM.

This commit uses the fma functions from math.h instead of relying on the compiler to decide to use them.