MAINT: add a 'tests' install tag to the `numpy._core._simd` extension module
This avoids installing `_simd.so` when the 'tests' install tag is omitted
in order to strip the whole test suite.
This is a significant saving in binary size especially on x86-64 (see
issue 25737, about 10% of the installed size).

Tested on macOS arm64 too, there `_simd.so` is 363 kb for a release build,
or ~12% of the size of the main extension module (`_multiarray_umath.so`).