BUG: Add object cast to avoid warning with limited API
To be honest, I think we should just delete this from the public API.
But I thought I'll start with this, since at least in
principle that isn't a bug-release thing to backport.

I am not sure setting warnings to errors is wise (at least for Cython and
MSVC).  OTOH, the Cython module currently truly does nothing except
include the headers (it doesn't even use the NumPy `.pyd` yet).