MAINT: back printoptions with a true context variable (#26846)
This is a re-do of gh-26345, I'm taking over from @mtsokol because this is needed for the free-threaded work.

The new _printoptions.py file exists to avoid a circular import during setup of the multiarray module.

I'm guessing this adds some overhead to printing. I haven't benchmarked it because it wasn't clear to me: do we care about printing performance? I could certainly add some caching or a way to avoid repeatedly calling get_legacy_print_mode for every printed value. We could also keep the C global we had before but make it thread-local. I just thought it made things conceptually simpler to store all the printoptions state in the context variable.

Co-authored-by: Mateusz Sokół <mat646@gmail.com>