BUG: Fix simd loadable stride logic
It was never correct that strides are divisable by their itemsize and
*some* code even checked for it correctly, while others had asserts
in place that had no reason why they should be guaranteed to pass.

This moves the check into the `loadable_stride` macro.  The exp
implementation has it's own custom logic, that is probably OK.

I assume that compilers will optimize the duplicate division away,
if that may not be the case, the code should be optimized to avoid
it probably.