BUG: random: In Johnk's algorithm, use logarithms if either X or Y is 0.
If either X or Y is zero, it means the calculation has underflowed
and information has been lost.  So use the logarithm-based calculation
if *either* is 0.  Before this, *both* had to be 0 for the logarithm-based
calculation to be used. If X was 0 and Y was not, the result X/(X + Y) would
be 0, when in fact the logarithm-based calculation might give the correct
small but nonzero value.

Closes gh-24475.