MAINT: linalg: Simplify some linalg gufuncs.
The new feature added in gh-26908 is used to reduce the number
of gufuncs in np.linalg._umath_linalg.  For example, before
this change, the gufuncs `svd_m` and `svd_n` used the same
code, but two distinct gufuncs were necessary because,
for input with shape (m, n), the output has shape min(m, n).
`svd_m` returned shape (m,), and `svd_n` returned shape (n,).
The wrapper code had to determine which gufunc to call.
This can be handled now with a single gufunc by defining
a function that sets the output dimension to min(m, n).

In this PR, the following pairs of gufuncs are coalesced into
a one gufunc:

    lstsq_m & lstsq_n       => lstsq
    qr_r_raw_m & qr_r_raw_n => qr_r_raw
    svd_m & svd_n           => svd
    svd_m_f & svd_n_f       => svd_f
    svd_m_s & svd_n_s       => svd_s