MAINT: Backport f2py from main.
There are enough changes to f2py in the main branch to make a wholesale
backport the easy way to go. There are a number of GIL related changes
which are irrelevant for the 2.0.x branch, but should be harmless. Same
for various style fixups, but disentangling what is relevent from what
is not isn't worth the effort.