API,BUG: Fix copyto (and ufunc) handling of scalar cast safety
This defines scalar cast safety to be "safe", "equiv", or "unsafe"
(for Python int, float, complex).
Although, equiv is a bit of an odd one.  We define it by using
the common DType with the desired one, and assuming that this cast
is at least safe.  With that, we can create an appropriate array
for the scalar (and the rest works fine).

This means that any python float can be safely cast to e.g. float32
and any Python integer to any NumPy integer even though the
actual assignment may overflow (floats) or fail (integers).
`can_cast` itself is not defined here.  It could be, but has the
disadvantage that it won't error out on assignment later for
integers.