DEP: lib: Deprecate acceptance of float (and more) in bincount. (#27076)
The first argument of bincount is expected to contain integers.
As noted in gh-3138, it actually accepts floats, and casts them
to integers with no warnings. (It also accepts other objects that
can be cast to integers, so inputs such as ["1", Fraction(5, 3)]
are accepted, with fractional parts silently dropped.)

This change deprecates that behavior.  Now a deprecation warning
is generated if the input cannot be safely cast to integer.

Closes gh-3138.

Co-authored-by: Sebastian Berg <sebastian@sipsolutions.net>