TYP: Make array shape type variable covariant and bound
Fixes #25729

This change allows future changes to the static typing of numpy that modify
or only work with certain numbers of dimensions.  It also applies the change
to subclasses of ndarray and adds tests.  It allows users to statically
type their array shapes with subtypes of tuple (e.g. NamedTuple) and tuples
of int subtypes (e.g. Literal or NewType).

For a discussion of the merits of TypeVarTuple vs a tuple-bound TypeVar, see
the linked PR