BUG: Ensure that scalar binops prioritize __array_ufunc__
If array-ufunc is implemented, we must call always use it for all
operators (that seems to be the promise).

If __array_function__ is defined we are in the clear w.r.t. recursion
because the object is either an array (can be unpacked, but already
checked earlier now also), or it cannot call the ufunc without unpacking
itself (otherwise it would cause recursion).

There is an oddity about `__array_wrap__`.  Rather than trying to
do odd things to deal with it, I added a comment explaining why it
doens't matter (roughly: don't use our scalar priority if you want
to be sure to get a chance).