MAINT: Remove any promotion-state switching logic
This is the first level, no following code simplifications, just
straight up deletions of any branching.

I kept a UserWarning, in case someone had the bad idea to permanently
set the environment variable and think they can rely on it.
Although would be happy to just delete that as well.