BUG: Disable SVE VQSort
This patch removes the SVE dispatch path for VQSort, due to it being broken with GCC 10.2.1 in the manylinux2014 image.

Compiling it outside of manylinux2014 with GCC 10.5.0 appears to work correctly.

I'm assuming this isn't being caught in CI due to there not being a SVE capable machine in the wheel builds?