BUG: Fix user dtype can-cast with python scalar during promotion
The can-cast code for "Python scalars" was old and did not correctly
take into account possible user-dtypes with respect to NEP 50 weak
promotion.

To do this, we already had the necessary helper functions that go
via promotion (although it took me some brooding to remember ;)).

So the fix is rather simple.  Actually adding CI/test for the fix
is unfortunately hard as it requires such a user DType.