Update basics.types.rst: Clarify determinant precision handling
This update adds clarification to the section on numerical data types regarding the precision of matrix determinants in NumPy. A note has been included explaining that results close to zero, such as those from the determinant of the matrix np.array([[5,5,6],[7,7,5],[4,4,8]]), can appear as very small negative numbers due to floating-point precision issues.

Additionally, a reference to the `np.isclose()` function has been added to guide users on checking if a value is close to zero, which can help mitigate confusion regarding the expected result of zero for certain matrix determinants.