DOC: fix name of shape parameter kappa of von Mises distribution (#27721)
* Fix name of shape parameter kappa of von Mises distribution

The parameter kappa in the equation of the von Mises distribution is not the dispersion but its reciprocal. This is typically referred to as the concentration (see for example https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.vonmises.html). Change the docstring of the `vonmises` method in `_generator.pyx` and `mtrand.pyx` accordingly.