BUG: Ensure that same-kind casting works for uints (mostly)
This is the most minimal fix I could think off... If we have uints
still use the out-of-bounds value, but make it uint.  This is because
the code uses copyto with the default same-kind casting in places...

This doesn't remove all the other quirks, and surely, it is a behavior
change in some awkward situations (since you got a uint, someone might
mix the newly uint value with an integer and get float64 results, etc.)

But... it is by far the most minimal fix I could think of after a longer
time.

A more thorough fix may be to just always store the exact dtype, but
it would propagate differently e.g. when casting.  A start for this
is currently in gh-27596.

Closes gh-27269, gh-27580