BUG,ENH: Fix internal ``__array_wrap__`` for direct calls (#27807)
* BUG,ENH: Fix internal ``__array_wrap__`` for direct calls

Since adding `return_scalar` as am argument, the array-wrap
implementations were slightly wrong when that argument was actually
passed and the function called directly.

NumPy itself rarely (or never) did so for our builtin types now
so that was not a problem within NumPy.

Further, the scalar version was completely broken, converting to
scalar even when such a conversion was impossible.

As explained in the code.  For array subclasses we NEVER want to
convert to scalar by default.  The subclass must make that choice
explicitly.
(There are plenty of tests for this behavior.)

* BUG: Ensure cast to self in ndarray.__array_wrap__ and other review fixes