MAINT: Refactor buffered loop to do an initial setup
This reorganizes the loop to do a larger initial setup rather than
figuring things out on every iteration.
That shrinks the buffersize often to fit better with the actual size.
We try to do that smartly, although, the logic is different.

It also means that "reduce" is not necessarily used (although it
is likely that before it was also effectively unused often).

This is not generally faster, but in some cases it will not use
buffering unnecessarily and it should be much better at buffer
re-use (at least with smaller fixes).
In those cases things will be much faster.

(As of writing, the setup time seems slightly slower, that is probably
not unexpected, since finding the best size is extra work and it would
only amortize if there are many iterations.)