MAINT: Ensure correct handling for very large unicode strings
In the future, we can handle these strings (in parts we already
can maybe), but for now have to stick to `int` length because
more of the code needs cleanup to actually use it safely.
(For user dtypes this is less of a problem, although corner cases
probably exist.)

This adds necessary checks to avoid large unicode dtypes.