MAINT: Move uint aligned check to actual transfer function setup
The comment was already always incorrect, all newly allocated
arrays are uint aligned.  But more importantly, we don't often
need casting, so we should move this check to where the casts are
required.

Since newly allocated arrays should not need casting or buffering
we shouldn't check things unnecessarily.

(While before, we checked things unnecessarily all the time.)

Locally, I saw a flaky linalg warning, but I think it is the one
we often saw (just surprised because I didn't see it in a long time).