MAINT: Move user pointers out of axisdata and simplify iternext
The pointers were always stored for each axis (pointing to the
start of the whatever the index into the axis was).
I found that confusing: One set of pointers is enough.

But some gufuncs relied on external loop giving a second set of
pointers (buffers always do so ufuncs are fine).
So, merged the buffer pointers and the external loop into a new
field.

I hope (but not sure) that compilers won't complain about
loops that are never taken (didn't notice it locally, but...).

For speed testing, I tried to time with something like:
```
a = np.broadcast_to(np.zeros((2, 4))[:, ::2], (100000, 2, 2))

%timeit np.count_nonzero(a)
%timeit np.count_nonzero(a[..., 0])

%timeit np.nonzero(a)
%timeit np.nonzero(a[..., 0])
```
Especially the first (count nonzero).  That code uses the external
loop path without buffering (i.e. what changed most).
Locally, saw a small slowdown with nonzero, but on a different box
and if anything I think a slight speedup for count-nonzero.

So, I doubt this makes a difference, of course an iteration only
C test might see a tiny bit more of a difference.