MAINT: Make qualname tests more specific and fix code where needed
This fixes the test to be more specific and more importantly not
unnecessarily skip things.

It also fixes two qualname errors (yes they are in almost private
things, but they should still be right).

Renamed the tests, since it is about qualname, but it is also about
the fact that you should be able to find any object using the module
and qualname.

Note, that the semi-private ufuncs inside `numpy._core` (skipped) would
not pass this tests because they still lack a module and qualname.
(e.g. `_clip`).