ENH: Allow an arbitrary number of operands in nditer
Internally, the change is very moderate mainly requiring to use
`int` in the struct (growing it a bit) and otherwise a temporary
work-array in the buffer setup function (one could probably use
scratch space on the nditer, but this seemed OK to me).

The big and somewhat annoying change is that the Python wrapper
needs a lot of per operand space, so setting it up is somewhat
annoying.