BUG: Ensure einsum uses chunking (now that nditer doesn't)
The nditer refactor enabled the `GROWINNER` for reductions, this
was not the case before (I am not sure about cases where the reduction
is in an outer axis, but I don't think so).

Either way, as the test says, chunking always improves the precision
of large sorts if their mean is nonzero and sklearn noticed lower
precision.

We could only remove growinner if there is a reduction, but it seems
like a 1% performance hit for the simplest (non-trivial) case where
GROWINNER had done something before.