TST: check that attributes in classes do not mask numpy functions
This test suite demonstrates that functions in
numpy/_core/fromnumerics.py wrongly assume that if an object has an
attribute with the same name of the function it is a callable method.

All test will fail with "TypeError: 'NoneType' object is not callable",
see GH28024.

A subsequent commit will fix these tests.