BUG: Fix building on s390x with clang
clang on s390x did not have implementations of vector logical
operators such as vec_and, vec_or and vec_xor in vecintrin.h
until __VEC__ == 10305 which caused compile errors. Add
implementations to allow the build to complete.

Currently, clang >= 19 is required for all tests to pass because
that is the minimum version supported by highway on s390x with clang.