MAINT: No need to check for check for FPEs in casts to/from object (#28358)
* MAINT: No need to check for check for FPEs in casts to/from object

Since these go via Python (in some form) and Python doesn't use FPEs
we can be sure that we don't need to check for FPEs.

Note that while it hides *almost always* spurious FPEs seen on some
platforms, there could be certain chains or multiple cast situations
where FPEs are checked for other reasons and the spurious FPE will
show up.

So it "somewhat":

Closes gh-28351

* MAINT: Follow-up, got the wrong place (the other is OK).

* DOC: Add a small comment as per review request

I don't think it needs a comment that we can do this, but maybe it
is nice to say that there was a reason for it.